pyproject.toml
[tool.poetry]
name = "k8s learn"
version = "0.1.0"
description = ""
authors = ["Valery Semenov <v.semenov@ylab.io>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.78.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.8.0"
SQLAlchemy = "^1.4.39"
databases = "^0.6.0"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
email-validator = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
